Multigrid Problems

Interface

FerriteMultigrid.AbstractPMultigrid — Type

abstract type AbstractPMultigrid

This is an abstract type that can be extended to when Rediscretization strategy is used as coarsening strategy, otherwise it's not required.

FerriteMultigrid.assemble — Function

assemble(problem::AbstractPMultigrid, fe_space::FESpace)

The interface that needs to be implemented to define the Rediscretization strategy for a specific problem type.

Diffusion

FerriteMultigrid.DiffusionMultigrid — Type

DiffusionMultigrid{C} <: AbstractPMultigrid

This struct represents a multigrid problem for diffusion equations with a coefficient C.

Linear Elasticity

FerriteMultigrid.LinearElasticityMultigrid — Type

LinearElasticityMultigrid{TC} <: AbstractPMultigrid

This struct represents a multigrid problem for linear elasticity equations with a material stiffness tensor `TC